+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import {Provider} from 'react-redux';
 import App from './app/app';
-import {store} from './store/index';
+import {store} from './store/store';
 import Error from './components/error/error';
 import {checkAuthAction, fetchFilmsAction, fetchPromoFilmAction} from './store/api-actions';
