+++ src/middlewares/redirect.ts
@@ -1,7 +1,7 @@
 import {PayloadAction} from '@reduxjs/toolkit';
-import browserHistory from '../../history-route/browser-history';
+import browserHistory from '../history-route/browser-history';
 import {Middleware} from 'redux';
-import {combineReducer} from '../combine-reducer';
+import {combineReducer} from '../store/combine-reducer';
 
 type Reducer = ReturnType<typeof combineReducer>
 export const redirect: Middleware<unknown, Reducer> =
@@ -9,7 +9,11 @@ export const redirect: Middleware<unknown, Reducer> =
     (next) =>
       (action: PayloadAction<string>) => {
         if (action.type === 'app/redirectToRoute') {
-          browserHistory.push(action.payload);
+          if (action.payload === '/back') {
+            browserHistory.back();
+          } else {
+            browserHistory.push(action.payload);
+          }
         }
         return next(action);
       };
