+++ src/store/film-part/film-actions/film-actions.ts
@@ -1,6 +1,6 @@
 import {createAction} from '@reduxjs/toolkit';
-import {Film, Films} from '../../types/film-type';
-import {Reviews} from '../../types/review-type';
+import {Film, Films} from '../../../types/film-type';
+import {Reviews} from '../../../types/review-type';
 
 export const loadSimilarFilms = createAction<Films>('data/loadSimilarFilms');
 export const loadFilm = createAction<Film>('data/loadFilm');
