+++ src/types/state.ts
@@ -1,4 +1,4 @@
-import {store} from '../store';
+import {store} from '../store/store';
 
 export type State = ReturnType<typeof store.getState>;
